very new

+++ myadmin/views.py
@@ -6,7 +6,7 @@ from accounts.models import Account
 from orders.models import Orders
 from theproducts.models import Product , Categoryies
 from django.views.decorators.cache import cache_control
-from django.core.paginator import Paginator
+from django.core.paginator import Paginator ,EmptyPage, InvalidPage
 
 
 # Create your views here.
@@ -50,7 +50,11 @@ def admin_home(request):
 @cache_control(no_cache =True, must_revalidate =True, no_store =True)
 def admin_table(request):
     if 'username' in request.session:
-        values = Account.objects.all()
+        value = Account.objects.all().order_by('id')
+        p = Paginator(value,9)
+        page = request.GET.get('page')
+        values = p.get_page(page)
+
         return render(request,'admintable.html',{'values':values})
 
     return redirect(adminLogin)
@@ -58,10 +62,10 @@ def admin_table(request):
 @cache_control(no_cache =True, must_revalidate =True, no_store =True)
 def productList(request):
     if 'username' in request.session:
-        values = Product.objects.all()
+        values = Product.objects.all().order_by('-id')
 
         #setup pagination
-        p = Paginator(Product.objects.all(),3)
+        p = Paginator(values,3)
         page = request.GET.get('page')
         productValues =p.get_page(page)
 
@@ -74,7 +78,7 @@ def productList(request):
 @cache_control(no_cache =True, must_revalidate =True, no_store =True)
 def categoryList(request):
     if 'username' in request.session:
-        values = Categoryies.objects.all()
+        values = Categoryies.objects.all().order_by('id')
         return render(request,'categories.html',{'values':values})
     return redirect(adminLogin)
 
@@ -92,6 +96,9 @@ def editproduct(request,id):
         product_stock = request.POST.get('stock')
         # product_category = request.POST.get('category')
         product_image = request.POST.get('image')
+        product_image1 = request.POST.get('image1')
+        product_image2 = request.POST.get('image2')
+
 
         obj = Product.objects.get(id=id)
 
@@ -101,6 +108,9 @@ def editproduct(request,id):
         obj.stock = product_stock
         # obj.category = product_category
         obj.image = product_image
+        obj.image1 = product_image1
+        obj.image2 = product_image2
+
 
         obj.save()
         return redirect(productList)
@@ -109,9 +119,12 @@ def editproduct(request,id):
        
 
 def deleteproduct(request,id):
-    my_product =Product.objects.get(id=id)
-    my_product.delete()
-    return redirect(productList)
+    if request.method == "POST":
+
+        print("tttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttttt")
+        my_product =Product.objects.get(id=id)
+        my_product.delete()
+        return redirect(productList)
 
 
 
@@ -126,8 +139,11 @@ def addproduct(request):
         
 
         product_image = request.POST.get('image')
+        product_image1 = request.POST.get('image1')
+        product_image2 = request.POST.get('image2')
+
         product = Product(name = product_name ,description =product_description,price=product_price,
-        stock= product_stock,image=product_image )
+        stock= product_stock,image=product_image,image1=product_image1,image2=product_image2 )
         product.category  = Categoryies.objects.get(id=categ)
         product.save()
         return redirect(productList)
@@ -141,8 +157,10 @@ def addcategory(request):
     if request.method == "POST":
         category_name = request.POST.get('category_name')
         description = request.POST.get('description')
+        offer = request.POST.get('offer')
+
         
-        cat = Categoryies(category_name=category_name,description=description)
+        cat = Categoryies(category_name=category_name,description=description,offer=offer)
         
         cat.save()
         return redirect(categoryList)
@@ -170,9 +188,12 @@ def blockuser(request,id):
 
 
 def orderdisplay(request):
-    order= Orders.objects.all().order_by('id')
-
-    return render(request,'orderadmin.html',{ 'order':order })
+    order= Orders.objects.all().order_by('-id')
+    p = Paginator(order,9)
+    page = request.GET.get('page')
+    orders = p.get_page(page)
+    
+    return render(request,'orderadmin.html',{ 'order':order,'orders':orders})
 
 def ordercanceladmin(request,id):
     
@@ -181,3 +202,45 @@ def ordercanceladmin(request,id):
     order.save()
 
     return redirect(orderdisplay)
+
+    
+
+def orderstatus(request,id):
+    if request.method == "POST":
+        status = request.POST.get('status')
+    
+        order = Orders.objects.get(id = id)
+        order.status = status
+        order.save()
+
+    return redirect(orderdisplay)
+def offerstatus(request,id):
+    if request.method == "POST":
+        status = request.POST.get('offer')
+    
+        categ = Categoryies.objects.get(id = id)
+        categ.offer = status
+        categ.save()
+
+    return redirect(categoryList)
+
+def  searchprod(request):
+    try:
+        q = request.GET['search']
+
+        data = Product.objects.all()
+        datas = []
+    
+        for i in data :
+            datas.append(i.name)
+
+        for i in datas:
+           if q.lower() in i.lower():
+               searched = Product.objects.filter(name = i)
+
+        print(searched)
+
+        return render(request, 'searcheditem.html',{'searched':searched})
+    except:
+        return render(request,'noproduct.html')
+
